rm gcs test util
