+++ src/ray/common/BUILD.bazel
@@ -11,14 +11,18 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "test_util",
-    srcs = ["test_util.cc"],
-    hdrs = ["test_util.h"],
+    name = "test_utils",
+    srcs = ["test_utils.cc"],
+    hdrs = ["test_utils.h"],
     deps = [
         ":asio",
         ":id",
         ":ray_object",
+        ":task_common",
+        "//src/ray/protobuf:autoscaler_cc_grpc",
         "//src/ray/protobuf:common_cc_proto",
+        "//src/ray/protobuf:gcs_cc_proto",
+        "//src/ray/protobuf:gcs_service_cc_grpc",
         "//src/ray/util:cmd_line_utils",
         "//src/ray/util:network_util",
         "//src/ray/util:path_utils",
