+++ src/ray/core_worker/task_execution/tests/concurrency_group_manager_test.cc
@@ -19,7 +19,7 @@
 
 #include "gtest/gtest.h"
 #include "ray/common/asio/instrumented_io_context.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/core_worker/task_execution/fiber.h"
 #include "ray/core_worker/task_execution/thread_pool.h"
