+++ src/ray/core_worker/task_submission/tests/BUILD.bazel
@@ -8,7 +8,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker/task_submission:dependency_resolver",
         "@com_google_googletest//:gtest",
     ],
@@ -47,7 +47,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/common:asio",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:actor_creator",
         "//src/ray/core_worker:reference_count",
         "//src/ray/core_worker:task_manager",
@@ -67,7 +67,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/fakes/ray/rpc/raylet:fake_raylet_client",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker/task_submission:normal_task_submitter",
         "//src/ray/raylet_client:raylet_client_lib",
