+++ src/ray/core_worker/tests/BUILD.bazel
@@ -32,7 +32,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/common:status",
         "//src/ray/common:status_or",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:memory_store",
         "@com_google_absl//absl/synchronization",
         "@com_google_googletest//:gtest",
@@ -71,7 +71,7 @@ ray_cc_test(
         "//src/fakes/ray/rpc/raylet:fake_raylet_client",
         "//src/mock/ray/pubsub:mock_publisher",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker:object_recovery_manager",
         "//src/ray/object_manager:object_manager_common",
@@ -91,7 +91,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/mock/ray/pubsub:mock_publisher",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker:reference_count",
         "//src/ray/core_worker:task_event_buffer",
@@ -110,7 +110,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:task_event_buffer",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/util:event",
@@ -133,7 +133,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:task_event_buffer",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/util:event",
@@ -152,7 +152,7 @@ ray_cc_test(
     tags = ["team:core"],
     deps = [
         "//:ray_mock",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:actor_creator",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/util:path_utils",
@@ -169,7 +169,7 @@ ray_cc_test(
     tags = ["team:core"],
     deps = [
         "//:ray_mock",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:common",
         "//src/ray/core_worker:generator_waiter",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
@@ -186,7 +186,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:actor_manager",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "@com_google_googletest//:gtest",
@@ -246,7 +246,7 @@ ray_cc_test(
         "//:ray_fakes",
         "//:ray_mock",
         "//src/fakes/ray/rpc/raylet:fake_raylet_client",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/core_worker:core_worker_lib",
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker:reference_count",
