+++ src/ray/core_worker/tests/actor_creator_test.cc
@@ -18,7 +18,7 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "ray/core_worker/actor_creator.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/util/path_utils.h"
 #include "ray/util/raii.h"
 #include "mock/ray/gcs/gcs_client/gcs_client.h"
