+++ src/ray/core_worker/tests/object_recovery_manager_test.cc
@@ -28,7 +28,7 @@
 #include "mock/ray/pubsub/publisher.h"
 #include "ray/common/task/task_spec.h"
 #include "ray/common/task/task_util.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/core_worker/store_provider/memory_store/memory_store.h"
 #include "ray/raylet_client/raylet_client.h"
