+++ src/ray/core_worker/tests/task_event_buffer_export_event_test.cc
@@ -28,7 +28,7 @@
 #include "gtest/gtest.h"
 #include "mock/ray/gcs/gcs_client/gcs_client.h"
 #include "ray/common/task/task_spec.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/core_worker/task_event_buffer.h"
 #include "ray/util/event.h"
