+++ src/ray/core_worker/tests/task_event_buffer_test.cc
@@ -34,7 +34,7 @@
 #include "mock/ray/gcs/gcs_client/gcs_client.h"
 #include "ray/common/task/task_spec.h"
 #include "ray/common/task/task_util.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/util/event.h"
 
 using ::testing::_;
