+++ src/ray/gcs/gcs_client/tests/BUILD.bazel
@@ -9,7 +9,7 @@ ray_cc_test(
     tags = ["team:core"],
     deps = [
         "//src/ray/gcs/gcs_client:gcs_client_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -33,7 +33,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/gcs/gcs_client:global_state_accessor_lib",
         "//src/ray/gcs/gcs_server:gcs_server_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:path_utils",
         "//src/ray/util:raii",
         "@com_google_googletest//:gtest_main",
@@ -62,7 +62,7 @@ ray_cc_test(
     deps = [
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/gcs/gcs_server:gcs_server_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:network_util",
         "//src/ray/util:raii",
         "@com_google_googletest//:gtest_main",
@@ -89,7 +89,7 @@ ray_cc_test(
     deps = [
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/gcs/gcs_server:gcs_server_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:network_util",
         "//src/ray/util:path_utils",
         "//src/ray/util:raii",
