+++ src/ray/gcs/gcs_server/tests/BUILD.bazel
@@ -21,7 +21,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/gcs/gcs_server:gcs_placement_group_manager",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:counter_map",
         "@com_google_googletest//:gtest_main",
     ],
@@ -48,7 +48,7 @@ ray_cc_test(
     ],
     deps = [
         "//src/ray/gcs/gcs_server:gcs_server_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -73,7 +73,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_store_client_kv",
         "//src/ray/gcs/store_client:in_memory_store_client",
         "//src/ray/gcs/store_client:redis_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -109,7 +109,7 @@ ray_cc_test(
         "//src/fakes/ray/rpc/raylet:fake_raylet_client",
         "//src/mock/ray/pubsub:mock_publisher",
         "//src/ray/gcs/gcs_server:gcs_node_manager",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -127,7 +127,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_job_manager",
         "//src/ray/gcs/gcs_server:gcs_kv_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -142,7 +142,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/gcs/gcs_server:gcs_task_manager",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -162,7 +162,7 @@ ray_cc_test(
         "//src/mock/ray/pubsub:mock_publisher",
         "//src/ray/gcs/gcs_server:gcs_placement_group_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:counter_map",
         "@com_google_googletest//:gtest_main",
     ],
@@ -188,7 +188,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_resource_manager",
         "//src/ray/gcs/gcs_server:gcs_table_storage",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:counter_map",
         "@com_google_googletest//:gtest_main",
     ],
@@ -205,12 +205,11 @@ ray_cc_test(
         "//src/fakes/ray/rpc/raylet:fake_raylet_client",
         "//src/fakes/ray/rpc/worker:fake_core_worker_client",
         "//src/mock/ray/pubsub:mock_publisher",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/gcs/gcs_server:gcs_actor",
         "//src/ray/gcs/gcs_server:gcs_actor_scheduler",
         "//src/ray/gcs/gcs_server:gcs_resource_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
         "//src/ray/util:counter_map",
         "@com_google_googletest//:gtest_main",
     ],
@@ -225,7 +224,7 @@ ray_cc_test(
     tags = ["team:core"],
     deps = [
         "//:ray_mock",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/gcs/gcs_server:gcs_actor",
         "//src/ray/gcs/gcs_server:gcs_actor_scheduler",
         "//src/ray/util:counter_map",
@@ -251,7 +250,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_actor_scheduler",
         "//src/ray/gcs/gcs_server:gcs_function_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/pubsub:publisher",
         "@com_google_googletest//:gtest_main",
     ],
@@ -269,7 +268,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_store_client_kv",
         "//src/ray/gcs/gcs_server:gcs_worker_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:process",
         "@com_google_googletest//:gtest_main",
     ],
@@ -304,7 +303,7 @@ ray_cc_test(
         ":gcs_table_storage_test_lib",
         "//src/ray/gcs/gcs_server:gcs_table_storage",
         "//src/ray/gcs/store_client/tests:store_client_test_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -316,11 +315,10 @@ ray_cc_test(
     tags = ["team:core"],
     deps = [
         ":gcs_table_storage_test_lib",
-        "//src/ray/common:test_util",
         "//src/ray/gcs/gcs_server:gcs_table_storage",
         "//src/ray/gcs/store_client:in_memory_store_client",
         "//src/ray/gcs/store_client/tests:store_client_test_lib",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -341,7 +339,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_init_data",
         "//src/ray/gcs/gcs_server:gcs_resource_manager",
         "//src/ray/gcs/gcs_server:gcs_store_client_kv",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/raylet/scheduling:cluster_resource_manager",
         "@com_google_googletest//:gtest_main",
     ],
@@ -358,7 +356,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/gcs/gcs_server:gcs_node_manager",
         "//src/ray/gcs/gcs_server:gcs_resource_manager",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/raylet/scheduling:cluster_resource_manager",
         "@com_google_googletest//:gtest_main",
     ],
@@ -375,7 +373,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/common:asio",
         "//src/ray/gcs/gcs_server:gcs_usage_stats_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -394,7 +392,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_job_manager",
         "//src/ray/gcs/gcs_server:gcs_kv_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -416,7 +414,7 @@ ray_cc_test(
         "//src/ray/gcs/gcs_server:gcs_actor_scheduler",
         "//src/ray/gcs/gcs_server:gcs_function_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/pubsub:publisher",
         "//src/ray/rpc:core_worker_client",
         "//src/ray/util:event",
@@ -437,7 +435,7 @@ ray_cc_test(
         "//src/mock/ray/pubsub:mock_publisher",
         "//src/ray/gcs/gcs_server:gcs_node_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
-        "//src/ray/gcs/tests:gcs_test_util_lib",
+        "//src/ray/common:test_utils",
         "//src/ray/util:string_utils",
         "@com_google_googletest//:gtest",
     ],
