+++ src/ray/gcs/gcs_server/tests/export_api/gcs_node_manager_export_event_test.cc
@@ -22,9 +22,9 @@
 
 #include "fakes/ray/rpc/raylet/raylet_client.h"
 #include "mock/ray/pubsub/publisher.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/gcs_server/gcs_node_manager.h"
 #include "ray/gcs/store_client/in_memory_store_client.h"
-#include "ray/gcs/tests/gcs_test_util.h"
 #include "ray/util/event.h"
 #include "ray/util/string_utils.h"
