+++ src/ray/gcs/gcs_server/tests/gcs_server_test_util.h
@@ -26,7 +26,7 @@
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/common/lease/lease.h"
 #include "ray/common/task/task_util.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/gcs_server/gcs_actor_manager.h"
 #include "ray/gcs/gcs_server/gcs_actor_scheduler.h"
 #include "ray/gcs/gcs_server/gcs_node_manager.h"
