+++ src/ray/gcs/gcs_server/tests/gcs_table_storage_test_base.h
@@ -19,9 +19,8 @@
 
 #include "gtest/gtest.h"
 #include "ray/common/id.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
-#include "ray/gcs/tests/gcs_test_util.h"
 
 namespace ray {
