+++ src/ray/gcs/gcs_server/tests/in_memory_gcs_table_storage_test.cc
@@ -16,7 +16,7 @@
 
 #include <memory>
 
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
 #include "ray/gcs/gcs_server/tests/gcs_table_storage_test_base.h"
 #include "ray/gcs/store_client/in_memory_store_client.h"
