+++ src/ray/gcs/gcs_server/tests/redis_gcs_table_storage_test.cc
@@ -15,7 +15,7 @@
 #include <memory>
 
 #include "gtest/gtest.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
 #include "ray/gcs/gcs_server/tests/gcs_table_storage_test_base.h"
 #include "ray/gcs/store_client/redis_store_client.h"
