+++ src/ray/gcs/store_client/tests/redis_async_context_test.cc
@@ -20,7 +20,7 @@
 
 #include "gtest/gtest.h"
 #include "ray/common/asio/instrumented_io_context.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/store_client/redis_context.h"
 #include "ray/util/logging.h"
 #include "ray/util/path_utils.h"
