+++ src/ray/gcs/store_client/tests/store_client_test_base.h
@@ -26,7 +26,7 @@
 #include "absl/container/flat_hash_map.h"
 #include "ray/common/asio/io_service_pool.h"
 #include "ray/common/id.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/gcs/store_client/store_client.h"
 #include "ray/util/logging.h"
 #include "src/ray/protobuf/gcs.pb.h"
