--- src/ray/gcs/tests/BUILD.bazel
@@ -1,14 +1,0 @@
-load("//bazel:ray.bzl", "ray_cc_library")
-
-ray_cc_library(
-    name = "gcs_test_util_lib",
-    hdrs = [
-        "gcs_test_util.h",
-    ],
-    deps = [
-        "//src/ray/common:test_util",
-        "//src/ray/gcs:gcs_pb_util",
-        "//src/ray/protobuf:autoscaler_cc_grpc",
-        "//src/ray/protobuf:gcs_service_cc_grpc",
-    ],
-)
