+++ src/ray/raylet/scheduling/tests/BUILD.bazel
@@ -11,7 +11,7 @@ ray_cc_test(
         "//:ray_mock",
         "//src/ray/common:lease",
         "//src/ray/common:ray_config",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/raylet/scheduling:cluster_resource_scheduler",
         "@com_google_googletest//:gtest_main",
@@ -58,7 +58,7 @@ ray_cc_test(
         "//src/ray/common:id",
         "//src/ray/common:lease",
         "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
+        "//src/ray/common:test_utils",
         "//src/ray/raylet:local_lease_manager",
         "//src/ray/raylet/scheduling:cluster_lease_manager",
         "//src/ray/raylet/scheduling:cluster_resource_scheduler",
