+++ src/ray/raylet/scheduling/tests/cluster_resource_scheduler_test.cc
@@ -25,7 +25,7 @@
 #include "gtest/gtest.h"
 #include "ray/common/ray_config.h"
 #include "ray/common/task/task_util.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 #include "ray/common/scheduling/resource_set.h"
 #include "ray/common/scheduling/scheduling_ids.h"
 #include "mock/ray/gcs/gcs_client/gcs_client.h"
