+++ src/ray/raylet/tests/lease_dependency_manager_test.cc
@@ -23,7 +23,7 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "mock/ray/object_manager/object_manager.h"
-#include "ray/common/test_util.h"
+#include "ray/common/test_utils.h"
 
 namespace ray {
